+++ spec/models/spree/payment_method/stripe_credit_card_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe Spree::Gateway::StripeGateway do
+describe Spree::PaymentMethod::StripeCreditCard do
   let(:secret_key) { 'key' }
   let(:email) { 'customer@example.com' }
   let(:source) { Spree::CreditCard.new }
